tweak location hook types in `preact` package
