+++ packages/wouter-preact/types/use-browser-location.d.ts
@@ -1,33 +1,4 @@
-/*
- * Foundation: useLocation and paths
- */
-
-export type Path = string;
-
-// the base useLocation hook type. Any custom hook (including the
-// default one) should inherit from it.
-export type BaseLocationHook = (
-  ...args: any[]
-) => [Path, (path: Path, ...args: any[]) => any];
-
-/*
- * Utility types that operate on hook
- */
-
-// Returns the type of the location tuple of the given hook.
-export type HookReturnValue<H extends BaseLocationHook> = ReturnType<H>;
-
-// Returns the type of the navigation options that hook's push function accepts.
-export type HookNavigationOptions<H extends BaseLocationHook> =
-  HookReturnValue<H>[1] extends (
-    path: Path,
-    options: infer R,
-    ...rest: any[]
-  ) => any
-    ? R extends { [k: string]: any }
-      ? R
-      : {}
-    : {};
+import { Path, HookReturnValue } from "./location-hook";
 
 type Primitive = string | number | bigint | boolean | null | undefined | symbol;
 export const useLocationProperty: <S extends Primitive>(
@@ -35,7 +6,10 @@ export const useLocationProperty: <S extends Primitive>(
   ssrFn?: () => S
 ) => S;
 
-export const useSearch: () => string;
+type SearchString = `?${string}` | "";
+export const useSearch: (options?: {
+  ssrSearch?: SearchString;
+}) => SearchString;
 
 export const usePathname: (options?: { ssrPath?: Path }) => Path;
 
@@ -54,7 +28,7 @@ export const navigate: (
 // It operates on current URL using History API, supports base path and can
 // navigate with `pushState` or `replaceState`.
 export type LocationHook = (options?: {
-  base?: Path;
+  ssrPath?: Path;
 }) => [Path, typeof navigate];
 
 export const useBrowserLocation: LocationHook;
