+++ packages/wouter-preact/types/use-hash-location.d.ts
@@ -1,8 +1,7 @@
-import { Path } from "./use-browser-location";
+import { Path } from "./location-hook";
 
 export function navigate<S = any>(to: Path, options?: { state: S }): void;
 
 export function useHashLocation(options?: {
-  base?: Path;
   ssrPath?: Path;
 }): [Path, typeof navigate];
